feat: add mergeArrays tests
