+++ test/mergeAndCompare.test.ts
@@ -60,57 +60,3 @@ test('Extend with custom concat arrays', () => {
   // const res2 = mergeAndCompare(concatArr, ['a'], ['b'])
   // expect(res2).toEqual( ['a', 'b'])
 })
-
-test('undefined object', () => {
-  const origin = {
-    pages: {
-      'aa': 'ttt',
-    },
-  }
-
-  const newData = {
-    pages: {
-      'aa': '1111',
-      'bb': '2222',
-    }
-  }
-
-  function convertTimestamps(originVal: any, targetVal: any, key: any) {
-    if (originVal !== undefined)
-      return targetVal
-  }
-
-  const res = mergeAndCompare(convertTimestamps, origin, newData)
-  expect(res as any).toEqual({ pages: { aa: "1111", bb: undefined } })
-})
-
-test('undefined array', () => {
-  function convertTimestamps(originVal: any, targetVal: any, key: any) {
-    if (originVal !== undefined)
-      return targetVal
-  }
-  const origin = {
-    date: [
-      {
-        'new': 'aaa',
-      }
-    ]
-  }
-  const target = {
-    date: [
-      {
-        'new': 'aaa',
-        'old': 'bbb',
-      }
-    ]
-  }
-  const res = mergeAndCompare(convertTimestamps, origin, target)
-  expect(res as any).toEqual({
-    date: [
-      {
-        'new': 'aaa',
-        'old': undefined,
-      }
-    ]
-  })
-})
